requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
numpy
scikit-learn
weaviate-client
python-dotenv
